zope.msg instead of dtml for email, test for 124 finished, except for error msg
The dtml for the notification mails is replaced by zope.msg templates, i18n now
possible. The test for 124 tests now that exceptions from MailHost.send are caught.

However, they are not yet printed to the error log and the test is not testing
whether they are printed to the errorlog, yet.